Rewrite
This is a massive rewrite that has changed a lot of the core functionality. It has added a statsd server for command/error/message stats. It has added a redis server(not currently working) and is going to add a postgres library driver. It has moved it into a better file structure and improved the layout of many files. This does not currently have the full amount of features and more is to come along with automated builds and tests